fix(review):  프로팩션 응답값 타입 매핑 오류 해결 (#177)

+++ src/main/java/com/lxp/aplus/review/infrastructure/persistence/CustomReviewRepositoryImpl.java
@@ -19,7 +19,7 @@ public class CustomReviewRepositoryImpl implements CustomReviewRepository {
                 .select(Projections.constructor(ReviewSummary.class,
                         reviews.courseId,
                         reviews.rating.avg(),
-                        reviews.id.count()
+                        reviews.id.count().intValue()
                 ))
                 .from(reviews)
                 .where(reviews.courseId.in(courseIds))
